Move internal.OverlayStack into internal/driver/common
